+++ test/Sentry.Tests/Internals/SentryStackTraceFactoryTests.cs
@@ -37,8 +37,18 @@ namespace Other.Tests.Internals
             var stackTrace = sut.Create();
 
             Assert.NotNull(stackTrace);
-            Assert.Equal(nameof(Create_NoExceptionAndAttachStackTraceOptionOn_CurrentStackTrace), stackTrace.Frames.Last().Function);
-            Assert.DoesNotContain(stackTrace.Frames, p => p.Function == nameof(SentryStackTraceFactory.CreateFrame));
+
+            Assert.Equal(
+                nameof(Create_NoExceptionAndAttachStackTraceOptionOn_CurrentStackTrace) + "()",
+                stackTrace.Frames.Last().Function
+            );
+
+            Assert.DoesNotContain(stackTrace.Frames, p =>
+                p.Function?.StartsWith(
+                    nameof(SentryStackTraceFactory.CreateFrame) + '(',
+                    StringComparison.Ordinal
+                ) == true
+            );
         }
 
         [Fact]
@@ -73,7 +83,7 @@ namespace Other.Tests.Internals
 
             var stackTrace = sut.Create(exception);
 
-            Assert.Equal(new StackTrace(exception, true).FrameCount, stackTrace.Frames.Count);
+            Assert.Equal(new StackTrace(exception, true).FrameCount, stackTrace?.Frames.Count);
         }
 
         [Fact]