+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## vNext
+
+* Include full method signature in stack frames. [#662](https://github.com/getsentry/sentry-dotnet/pull/662)
+
 ## 3.0.0-alpha.7
 
 * Ref moved SentryId from namespace Sentry.Protocol to Sentry (#643) @lucas-zimerman